Add identifiers to the report metadata section
This patch adds 2 identifiers to the report metadata section:
- a scan id, which which allows to identify a specific scan run
- a URL to ensure we know where the report was generated from

The main use case behind this patch is to be able to use the reports
without having to add extra context. For now we agreed on a convention
to name the report file as `{owner}-{repo}-{timestamp}.json`, but this
information gets lost when the content of the repost is simply shared
via email or pastebin.